Added support for SOCKS4 proxies
It is now possible to import trixy.proxy and use the SOCKS4 implementation.
However, the proxy does not currently have support for the BIND command.

While testing this code, I also discovered a mistake in the main Trixy code
where data read by inputs would be directly sent to downstream notes instead
of being handled by the input node (which is needed when the node does some
type of protocol parsing, as the SOCKS nodes do). This was corrected.